requirements.txt
streamlit
crewai
langchain_community
python-dotenv
firebase-admin
